runtimes/huggingface/pyproject.toml
[tool.poetry]
name = "mlserver-huggingface"
version = "2.0.14"
description = "HuggingFace runtime for MLServer"
authors = ["Seldon Technologies Ltd. <hello@seldon.io>"]
license = "Apache-2.0"
readme = "README.md"
packages = [{include = "mlserver_huggingface"}]

[tool.poetry.dependencies]
python = ">=3.9,<3.12"
mlserver = "*"
tensorflow = "*"
pillow = "*"
optimum = {extras = ["onnxruntime"], version = ">=1.4,<2.0"}
pydantic = "2.7.1"
accelerate = "^0.27.2"
bitsandbytes = "^0.42.0"
sentence-transformers = "2.5.1"
transformers = "^4.48.0"
timm ="^1.0.13"

[tool.poetry.group.dev.dependencies]
mlserver = {path = "../..", develop = true}
# macOS x86 deprecated
# see: https://github.com/pytorch/pytorch/issues/114602
torch = [
    {markers = "sys_platform == 'darwin' and platform_machine == 'x86_64'", version = "<2.3.0"},
    {markers = "sys_platform != 'darwin' or platform_machine != 'x86_64'", version = "<2.4.0"}
]
transformers = ">=4.30,<5.0"
tensorflow = ">=2.12,<2.17"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
